- gtest for epoch mismatch (#20873)
- Fix p4ctl to make it functional on real H/W for Athena pipeline
- eth_dbgtool support for Athena builds
- Add static flow in AthenaApp
- Prelim support for modes in AthenaApp
        - Manually verified L2FWD mode
        - Needs to be supported via a config file/command line

+++ nic/apollo/api/include/athena/pds_epoch.h
@@ -24,12 +24,12 @@ extern "C" {
 
 /// \brief Epoch key
 typedef struct pds_epoch_key_s {
-    uint16_t    epoch_id;    ///< Epoch id
+    uint32_t    epoch_id;    ///< Epoch id
 } __PACK__ pds_epoch_key_t;
 
 /// \brief Epoch data
 typedef struct pds_epoch_data_s {
-    uint32_t    epoch;    ///< Epoch value
+    uint16_t    epoch;    ///< Epoch value
 } __PACK__ pds_epoch_data_t;
 
 /// \brief Epoch specification
